+++ client/internal/client/struct.go
@@ -3,6 +3,7 @@ package client
 import (
 	"github.com/pkg/errors"
 	"google.golang.org/grpc"
+	"google.golang.org/grpc/credentials"
 	"google.golang.org/grpc/credentials/insecure"
 
 	"github.com/valek177/auth/grpc/pkg/auth_v1"
@@ -48,13 +49,13 @@ func (c *ChatV1Client) Close() {
 }
 
 func NewAuthV1Client() (*AuthV1Client, error) {
-	// creds, err := credentials.NewClientTLSFromFile("/home/valek/microservices_course/auth/tls/service.pem", "")
-	// if err != nil {
-	// 	return nil, err
-	// }
+	creds, err := credentials.NewClientTLSFromFile("/home/valek/microservices_course/auth/tls/service.pem", "")
+	if err != nil {
+		return nil, err
+	}
 	conn, err := grpc.NewClient(
 		authAddress,
-		grpc.WithTransportCredentials(insecure.NewCredentials()),
+		grpc.WithTransportCredentials(creds),
 	)
 	if err != nil {
 		return nil, errors.WithStack(err)
